+++ contents/list/containers/__init__.py
@@ -1,5 +1,2 @@
-from contents.list.containers.elements.flatbutton import FlatButton
-from contents.list.containers.elements.listbutton import ListButton
-from contents.list.containers.elements.allproductbutton import AllProductButton
+from contents.list.containers.elements import *
 from contents.list.containers.visiblelistoflists import VisibleListOfLists
-from contents.list.containers.elements.addlist import AddList
