+++ contents/list/containers/elements.kv
@@ -1,0 +1,43 @@
+<AddList>:
+    size_hint: (.8, .15)
+    BoxLayout:
+        orientation: 'vertical'
+        TextInput:
+            color: (0,0,0,.1)
+            id: text_inp
+            on_touch_down: root.clear_text_inp()
+            text: 'Введите имя списка'.encode('cp1251').decode('utf-8')
+        FlatButton:
+            on_release: root.add_list_close_callback()
+            text: 'Ok'
+
+<AllProductSettingTextInputPrice>:
+    color: (0,0,0,.3)
+    on_touch_down: root.click_text_inp()
+    text: 'Цена'.encode('cp1251').decode('utf-8')
+
+<AllProductSettingTextInputQty>:
+    color: (0,0,0,.3)
+    on_touch_down: root.click_text_inp
+    text: 'Количество'.encode('cp1251').decode('utf-8')
+
+<AllProductButton>:
+    background_color: (1,1,1,1)
+    background_down: ''
+    background_normal: ''
+
+<FlatButton@ButtonBehavior+Label>:
+    canvas.before:
+        Color:
+            rgba: (1,.42,.05,1)
+        Rectangle:
+            pos: self.pos
+            size: self.size
+
+<ListButton@ButtonBehavior+Label>:
+    canvas.before:
+        Color:
+            rgba: (1,1,1,1)
+        Rectangle:
+            pos: self.pos
+            size: self.size
