--- contents/list/containers/elements/allproductbutton.kv
@@ -1,4 +1,0 @@
-<AllProductButton>:
-    background_color: (1,1,1,1)
-    background_down: ''
-    background_normal: ''
