--- contents/list/containers/elements/allproductsettingtextinputprice.kv
@@ -1,4 +1,0 @@
-<AllProductSettingTextInputPrice>:
-    color: (0,0,0,.3)
-    on_touch_down: root.click_text_inp()
-    text: 'Цена'.encode('cp1251').decode('utf-8')
