--- contents/list/containers/elements/allproductsettingtextinputqty.kv
@@ -1,4 +1,0 @@
-<AllProductSettingTextInputQty>:
-    color: (0,0,0,.3)
-    on_touch_down: root.click_text_inp
-    text: 'Количество'.encode('cp1251').decode('utf-8')
