--- contents/list/containers/elements/flatbutton.kv
@@ -1,7 +1,0 @@
-<FlatButton@ButtonBehavior+Label>:
-    canvas.before:
-        Color:
-            rgba: (1,.42,.05,1)
-        Rectangle:
-            pos: self.pos
-            size: self.size
