--- contents/list/containers/elements/listbutton.kv
@@ -1,7 +1,0 @@
-<ListButton@ButtonBehavior+Label>:
-    canvas.before:
-        Color:
-            rgba: (1,1,1,1)
-        Rectangle:
-            pos: self.pos
-            size: self.size
